requirements.txt
# Streaming & Messaging
confluent-kafka>=2.1.0,<2.9.0

# Search & Analytics
elasticsearch>=7.0.0,<=8.17.1

# Data Generation
faker>=30.0.0,<=35.0.0

# AWS SDK
boto3>=1.35.0,<=1.36.6
botocore>=1.35.0,<=1.36.26

# Airflow
apache-airflow==2.10.4